fix: Report `API` request on some file.
